@@ -74,56 +74,57 @@
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^0.26.0",
-    "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
-    "@iconify-json/ant-design": "^1.1.3",
-    "@iconify-json/bi": "^1.1.6",
-    "@iconify-json/cil": "^1.1.2",
-    "@iconify-json/clarity": "^1.1.4",
+    "@antfu/eslint-config": "^0.36.0",
+    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
+    "@iconify-json/ant-design": "^1.1.5",
+    "@iconify-json/bi": "^1.1.15",
+    "@iconify-json/cil": "^1.1.4",
+    "@iconify-json/clarity": "^1.1.8",
     "@iconify-json/eva": "^1.1.2",
-    "@iconify-json/ic": "^1.1.7",
-    "@iconify-json/la": "^1.1.2",
-    "@iconify-json/logos": "^1.1.14",
-    "@iconify-json/lucide": "^1.1.36",
-    "@iconify-json/material-symbols": "^1.1.8",
-    "@iconify-json/mdi": "^1.1.25",
-    "@iconify-json/mi": "^1.1.2",
-    "@iconify-json/ph": "^1.1.2",
-    "@iconify-json/ri": "^1.1.3",
-    "@iconify-json/simple-icons": "^1.1.29",
-    "@iconify-json/system-uicons": "^1.1.4",
-    "@iconify-json/vscode-icons": "^1.1.14",
-    "@intlify/vite-plugin-vue-i18n": "^6.0.1",
+    "@iconify-json/ic": "^1.1.13",
+    "@iconify-json/la": "^1.1.4",
+    "@iconify-json/logos": "^1.1.23",
+    "@iconify-json/lucide": "^1.1.70",
+    "@iconify-json/material-symbols": "^1.1.34",
+    "@iconify-json/mdi": "^1.1.50",
+    "@iconify-json/mi": "^1.1.4",
+    "@iconify-json/ph": "^1.1.5",
+    "@iconify-json/ri": "^1.1.5",
+    "@iconify-json/simple-icons": "^1.1.45",
+    "@iconify-json/system-uicons": "^1.1.8",
+    "@iconify-json/vscode-icons": "^1.1.22",
+    "@intlify/unplugin-vue-i18n": "^0.9.2",
     "@nuxt/image-edge": "^1.0.0-27657146.da85542",
+    "@pinia/nuxt": "^0.4.7",
     "@types/axios": "^0.14.0",
     "@types/dagre": "^0.7.48",
     "@types/file-saver": "^2.0.5",
     "@types/leaflet": "^1.9.0",
     "@types/leaflet.markercluster": "^1.5.1",
-    "@types/papaparse": "^5.3.2",
+    "@types/papaparse": "^5.3.7",
     "@types/sortablejs": "^1.13.0",
     "@types/tinycolor2": "^1.4.3",
     "@types/validator": "^13.7.10",
-    "@vitest/ui": "^0.18.0",
-    "@vue/compiler-sfc": "^3.2.37",
-    "@vue/test-utils": "^2.0.2",
-    "@vueuse/nuxt": "^9.0.2",
+    "@vitest/ui": "^0.29.2",
+    "@vue/test-utils": "^2.3.1",
+    "@vueuse/nuxt": "^9.13.0",
     "@windicss/plugin-animations": "^1.0.9",
     "@windicss/plugin-question-mark": "^0.1.1",
     "@windicss/plugin-scrollbar": "^1.2.3",
-    "eslint": "^8.22.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "happy-dom": "^6.0.3",
-    "nuxt": "^3.0.0-rc.12",
-    "nuxt-windicss": "^2.5.0",
-    "prettier": "^2.7.1",
-    "sass": "^1.53.0",
-    "unplugin-icons": "^0.14.7",
-    "unplugin-vue-components": "^0.22.4",
+    "eslint": "^8.36.0",
+    "eslint-config-prettier": "^8.7.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "esno": "^0.16.3",
+    "happy-dom": "^8.9.0",
+    "nuxt": "^3.3.1",
+    "nuxt-windicss": "^2.6.0",
+    "prettier": "^2.8.4",
+    "sass": "^1.59.2",
+    "unplugin-icons": "^0.15.3",
+    "unplugin-vue-components": "^0.24.1",
     "vite-plugin-monaco-editor": "^1.1.0",
-    "vite-plugin-purge-icons": "^0.9.0",
-    "vitest": "^0.18.0",
+    "vite-plugin-purge-icons": "^0.9.2",
+    "vitest": "^0.29.2",
     "windicss": "^3.5.6"
   },
   "overrides": {
